Fix build warning
git-svn-id: svn://svn.xara.com/Trunk/XaraLX@1154 c9fdbc50-1c0f-0410-8134-ac6df33ab1bd

+++ wxOil/menucmds.cpp
@@ -294,7 +294,7 @@ void FileNewTemplateAction(INT32 iNumberOfTemplate)
 		}
 		FileUtil::StopFindingFiles();
 
-		if( iNumberOfTemplate >= setSortFilename.size() )
+		if( iNumberOfTemplate >= (INT32)setSortFilename.size() )
 		{
 			//We failed to find a template. So remove this
 			//item from the menu.
